+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/AtomicQueryTest.java
@@ -21,7 +21,7 @@ package ai.grakn.test.graql.reasoner;
 import ai.grakn.GraknGraph;
 import ai.grakn.concept.Concept;
 import ai.grakn.exception.GraqlQueryException;
-import ai.grakn.graphs.AdmissionsGraph;
+import ai.grakn.test.graphs.AdmissionsGraph;
 import ai.grakn.test.graphs.CWGraph;
 import ai.grakn.graphs.GeoGraph;
 import ai.grakn.graql.Graql;
