+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/inference/AdmissionsInferenceTest.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.test.graql.reasoner.inference;
 
-import ai.grakn.graphs.AdmissionsGraph;
+import ai.grakn.test.graphs.AdmissionsGraph;
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.QueryBuilder;
 import ai.grakn.test.GraphContext;
